behaviours no longer singletons
This commit makes the server side of the behaviour use classmethods and makes behaviours into instances on the client side (previously both sides used singleton-instances, a rather convoluted approach). Also, now when behaviours are addded to the client driver, their instance is added, not a string representing them. This will allow behaviour-instance to become smarter on the client side in the future.